wait properly, only kill if alive, better colors

+++ main.go
@@ -8,9 +8,9 @@ import (
 	"os"
 	"os/exec"
 	"os/signal"
-	"strings"
 	"syscall"
 
+	"github.com/alessio/shellescape"
 	"github.com/fatih/color"
 )
 
@@ -18,18 +18,17 @@ var clear = flag.Bool("clear", false, "clear the terminal each time")
 var cmdParts []string
 
 var cmd *exec.Cmd
+var ready = make(chan struct{}, 1)
 
-func killIfRunning(cmd *exec.Cmd) {
-	if cmd != nil {
+func kill() {
+	if cmd != nil && cmd.ProcessState == nil {
 		syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
 	}
+	<-ready
 }
 
-func runCommand() {
-	blue := color.New(color.FgBlue)
-	reset := color.New(color.Reset)
-
-	killIfRunning(cmd)
+func spawn() {
+	kill()
 
 	cmdParts := flag.Args()
 	if *clear {
@@ -43,16 +42,22 @@ func runCommand() {
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
-	blue.Println("Running " + reset.Sprint(strings.Join(cmd.Args, " ")))
+	shellCommand := shellescape.QuoteCommand(cmd.Args)
+
+	fmt.Println(color.BlueString("trigger") + " " + color.CyanString(shellCommand))
 	if err := cmd.Start(); err != nil {
 		fmt.Fprintln(os.Stderr, err)
 	}
 	go func() {
-		if err := cmd.Wait(); err == nil {
-			blue.Println("Done")
+		err := cmd.Wait()
+		fmt.Print(color.BlueString("trigger"))
+		if err != nil {
+			fmt.Print(" " + color.RedString(err.Error()))
 		} else {
-			blue.Println("Done: " + err.Error())
+			fmt.Print(" " + color.BlueString("exit code 0"))
 		}
+		fmt.Println()
+		ready <- struct{}{}
 	}()
 }
 
@@ -60,27 +65,24 @@ func main() {
 	flag.Parse()
 	cmdParts = flag.Args()
 
+	ready <- struct{}{}
+
+	// Clean up on Ctrl+C
 	c := make(chan os.Signal, 1)
 	signal.Notify(c, os.Interrupt)
 	go func() {
-		for range c {
-			if cmd != nil {
-				syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-			}
-			os.Exit(0)
-		}
+		<-c
+		kill()
+		os.Exit(0)
 	}()
 
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/" {
-			return
-		}
-		runCommand()
+		spawn()
 	})
-	runCommand()
+	spawn()
 	err := http.ListenAndServe(":7416", nil)
 	if err != nil {
-		killIfRunning(cmd)
+		kill()
 		log.Fatal("ListenAndServe: ", err)
 	}
 }
